SalesSlipScanner/requirements.txt
openai
Pillow>=8.0
